@@ -4,9 +4,10 @@
     "description": "Launcher for Minecraft",
     "main": "build/main/index.js",
     "scripts": {
-        "dev": "npm run build:main && concurrently \"webpack serve\" \"npm:start:dev\" ",
-        "build": "npm run build:main && npm run build:renderer",
+        "dev": "concurrently \"npm:build:main\" \"npm:build:renderer\" && concurrently \"webpack serve\" \"npm:start:dev\" -k",
+        "build": "npm run build:main && npm run build:preload && npm run build:renderer",
         "build:main": "esbuild src/main/index.ts --platform=node --bundle --outdir=build/main --sourcemap=inline --external:electron",
+        "build:preload": "esbuild src/preload/index.ts --platform=node --bundle --outdir=build/preload --sourcemap=inline --external:electron",
         "build:renderer": "webpack",
         "start:dev": "cross-env DEV=true electron .",
         "start:prod": "electron .",
@@ -27,43 +28,43 @@
     },
     "homepage": "https://github.com/AuroraTeam/Launcher#readme",
     "devDependencies": {
-        "@sweetalert2/theme-dark": "^5.0.2",
+        "@sweetalert2/theme-dark": "^5.0.5",
         "@types/semver": "^7.3.8",
         "@types/ws": "^7.4.7",
-        "aurora-api": "^0.1.3",
+        "aurora-api": "^0.2.0-beta",
         "concurrently": "^6.2.1",
         "copy-webpack-plugin": "^9.0.1",
         "cross-env": "^7.0.3",
-        "css-loader": "^6.2.0",
-        "electron": "^13.2.2",
+        "css-loader": "^6.3.0",
+        "electron": "^13.4.0",
         "electron-devtools-installer": "^3.2.0",
-        "electron-packager": "^15.3.0",
+        "electron-packager": "^15.4.0",
         "file-loader": "^6.2.0",
-        "html-minimizer-webpack-plugin": "^3.1.1",
+        "html-minimizer-webpack-plugin": "^3.2.0",
         "less": "^4.1.1",
         "less-loader": "^10.0.1",
         "rimraf": "^3.0.2",
-        "sass": "^1.38.1",
+        "sass": "^1.42.1",
         "sass-loader": "^12.1.0",
-        "style-loader": "^3.2.1",
+        "style-loader": "^3.3.0",
         "stylus": "^0.54.8",
         "stylus-loader": "^6.1.0",
-        "sweetalert2": "^11.1.4",
-        "ts-loader": "^9.2.5",
-        "typescript": "^4.3.5",
+        "sweetalert2": "^11.1.7",
+        "ts-loader": "^9.2.6",
+        "typescript": "^4.4.3",
         "vue": "^2.6.14",
         "vue-loader": "^15.9.8",
         "vue-router": "^3.5.2",
         "vue-sweetalert2": "^5.0.2",
         "vue-template-compiler": "^2.6.14",
-        "webpack": "^5.51.1",
+        "webpack": "^5.54.0",
         "webpack-cli": "^4.8.0"
     },
     "dependencies": {
-        "esbuild": "^0.12.24",
+        "esbuild": "^0.12.29",
         "p-map": "^5.1.0",
         "semver": "^7.3.5",
-        "source-map-support": "^0.5.19",
-        "webpack-dev-server": "^4.0.0"
+        "source-map-support": "^0.5.20",
+        "webpack-dev-server": "^4.3.0"
     }
 }
