@@ -1,27 +1,21 @@
 import { AuroraAPI } from 'aurora-api';
+import { ConfigHelper } from 'main/helpers/ConfigHelper';
 import { App } from '..';
-const apiConfig = require('../../../config.json').api;
 
 export default class APIManager {
-    api: AuroraAPI | null = null;
+    api = new AuroraAPI(
+        ConfigHelper.getConfig().api.ws || 'ws://localhost:1370'
+    );
 
     constructor() {
-        this.connect();
-    }
-
-    async connect() {
-        try {
-            this.api = (await new AuroraAPI().connect(
-                apiConfig.url || 'ws://localhost:1370'
-            )) as AuroraAPI;
-            App.window.sendEvent('apiConnectSuccess');
-        } catch (error) {
+        this.api.onOpen = () => App.window.sendEvent('apiConnectSuccess');
+        this.api.onError = e => {
             App.window.sendEvent('apiConnectError', 'Ошибка при подключении');
-            console.log(error);
-        }
+            console.error(e);
+        };
     }
 
     public async send(type: string, data?: object) {
-        return await this.api?.send(type, data);
+        return await this.api.send(type, data);
     }
 }
