@@ -1,7 +1,7 @@
-import * as fs from "fs"
-import * as http from "http"
-import * as https from "https"
-import { URL } from "url"
+import * as fs from 'fs';
+import * as http from 'http';
+import * as https from 'https';
+import { URL } from 'url';
 
 export class HttpHelper {
     /**
@@ -12,20 +12,20 @@ export class HttpHelper {
      */
     static downloadFile(url: URL, path: string): Promise<boolean> {
         return new Promise((resolve, reject) => {
-            const handler = url.protocol === "https:" ? https : http
-            const tempFile = fs.createWriteStream(path)
-            tempFile.on("close", () => {
-                resolve(true)
-            })
+            const handler = url.protocol === 'https:' ? https : http;
+            const tempFile = fs.createWriteStream(path);
+            tempFile.on('close', () => {
+                resolve(true);
+            });
 
             handler
-                .get(url, (res) => {
-                    res.pipe(tempFile)
+                .get(url, res => {
+                    res.pipe(tempFile);
                 })
-                .on("error", (err) => {
-                    fs.unlinkSync(path)
-                    reject(err)
-                })
-        })
+                .on('error', err => {
+                    fs.unlinkSync(path);
+                    reject(err);
+                });
+        });
     }
 }
