@@ -1,21 +1,24 @@
-import { Response } from "aurora-api"
-import { ipcMain, IpcMainInvokeEvent } from "electron"
-import { App } from "../.."
+import { Response, ResponseError } from 'aurora-api';
+import { ipcMain, IpcMainInvokeEvent } from 'electron';
+import { App } from '../..';
 
 export default class LauncherAuth {
     constructor() {
-        ipcMain.handle('auth', this.auth)
+        ipcMain.handle('auth', this.auth);
     }
 
     async auth(_e: IpcMainInvokeEvent, login: string, password: string) {
         try {
-            const { data } = await App.api?.send('auth', {login, password}) as Response
-            return data
+            const { data } = <Response>(
+                await App.api.send('auth', { login, password })
+            );
+            return data;
         } catch (error) {
-            console.log(`Ошибка №${error.code}: ${error.message}`)
+            const e = <ResponseError>error; // Какого хрена TS?!
+            console.log(`Ошибка №${e.code}: ${e.message}`);
             return {
-                error: error.message
-            }
+                error: e.message
+            };
         }
     }
 }
