@@ -1,21 +1,26 @@
-import { Response } from "aurora-api"
-import { ipcMain, IpcMainInvokeEvent } from "electron"
-import { App } from "../.."
+import { Response, ResponseError } from 'aurora-api';
+import { ipcMain, IpcMainInvokeEvent } from 'electron';
+import { App } from '../..';
 
 export default class ServerList {
     constructor() {
-        ipcMain.handle('getServers', this.getServers)
+        ipcMain.handle('getServers', this.getServers);
     }
 
     async getServers(_e: IpcMainInvokeEvent) {
         try {
-            const { data } = await App.api?.api?.send('servers') as Response & { data: { servers: any[] } }
-            return data.servers
+            const { data } = <ServerResponse>await App.api.send('servers');
+            return data.servers;
         } catch (error) {
+            const e = <ResponseError>error; // Какого хрена TS?!
             return {
-                code: error.code,
-                message: error.message
-            }
+                code: e.code,
+                message: e.message
+            };
         }
     }
+}
+
+interface ServerResponse extends Response {
+    data: { servers: any[] };
 }
