@@ -1,21 +1,28 @@
-import { Response } from "aurora-api"
-import { ipcMain, IpcMainInvokeEvent } from "electron"
-import { App } from "../.."
+import { Response, ResponseError } from 'aurora-api';
+import { ipcMain, IpcMainInvokeEvent } from 'electron';
+import { App } from '../..';
 
 export default class ServerPanel {
     constructor() {
-        ipcMain.handle('getProfile', this.getProfile)
+        ipcMain.handle('getProfile', this.getProfile);
     }
 
     async getProfile(_e: IpcMainInvokeEvent, uuid: string) {
         try {
-            const { data } = await App.api?.api?.send('profile', {uuid}) as Response & { data: { profile: any[] } }
-            return data.profile
+            const { data } = <ProfileResponse>(
+                await App.api.send('profile', { uuid })
+            );
+            return data.profile;
         } catch (error) {
+            const e = <ResponseError>error; // Какого хрена TS?!
             return {
-                code: error.code,
-                message: error.message
-            }
+                code: e.code,
+                message: e.message
+            };
         }
     }
+}
+
+interface ProfileResponse extends Response {
+    data: { profile: any[] };
 }
